+++ print_uint.h
@@ -3,4 +3,4 @@
 #include <stdint.h>
 #include <stddef.h>
 
-int print_uint(char *out, uint64_t x);
+size_t print_uint(char *out, uint64_t x);
